CMake updates based on patches from Orcan Ogetbil: libinstpatch Python support, fixes to lib prefix variable with Fedora builds, CMakeLists.txt in utils/. pkg-config .pc file now installed also for MINGW build. Added RUNTIME install in libinstpatch/libinstpatch/CMakeLists.txt since it was needed for MinGW build (DLLs are installed in RUNTIME dir). Potential crash bug fix in ipatch_gig_file_identify_method(). Added REQUIRED to libinstpatch pkg_check_modules in swami/CMakeLists.txt.

+++ libinstpatch/libinstpatch/IpatchGigFile.c
@@ -57,7 +57,7 @@ ipatch_gig_file_identify_method (IpatchFile *file, IpatchFileHandle *handle,
                                  GError **err)
 {
   char *filename;
-  guint8 buf[3];
+  guint32 buf[3];
   gboolean retval = TRUE;
   int len;
